Changed a lot of things and made it to break absolutely nothing.
Now the API for the szena valuation process is much simple. Now you have a job function that needs the IDN structure and the data for the valuation.

The aggregation is able to get custom keys to label the scenarios, this is provided via the broker key, that now copies the cache variable of the Palm message if it is available. This is a very powerful trick, and it will be used to classify the scenarios in a wide variety of ways. For Szena, the pipeline and the session are the same thing, the same way that the cache and the scenario are too.

The next step is to make the whole system aware of the different scenarios, and support grouping in the aggregation. This will require important changes in the IDN script specification.