Make gateway router and dealer more generic.
The previous implementations were only useful for an http gateway. The goal here is to build more generic gateways to support multiple clients connected at once.

One important change is that from now on, the clients must specify their id as the socket identity to be able to recv the messages from the gateway.