This is a combination of 7 commits. Now the whole library needs only one client.
Tested with the trivial cases. Lots of cleaning up still to do.

Version bump to make explicit that lots of things changed.

The most important examples are now fixed

Imporoved logging messages of components and servers

Fixed the http worker example too.

Adapted logging example, and solved a regression.

A previous change had caused a regression in how the
cache storage object was passed to the different parts
of the master server. Now it is passed as a keyword
argument visible to the implementation, and it was
hidden previously.

Fixed cache example too